[nat64] add support for port translation
This commit adds support for port translation in NAT64.
Currently the translator only supports a direct mapping between an IPv4 and an IPv6 address
requiring a large IPv4 pool to work effectively. This implies the need for an additional
NAT 4 to 4 translator as the border router will only have one private IPv4 in the LAN.

This commit enhances the current functionality according to NAT64 RFC and dynamic PAT
mechanism. This is used by any home router to share a public IPv4 address with a range
of devices using private addresses. The port translation ensures the source port or the
ICMP identifier is always unique in the mapping table allowing a reply packet to be matched
with the request using the destination port and address. This functionality can be disabled
at compile time.

Signed-off-by: Marius Preda <marius.preda@nxp.com>